adding spaces in process
